all tests inherit HeadlessTestCase
